code: modify sources.schema.json and sources.csv files
