+++ technologydata/datasources/example03/sources.csv
@@ -1,0 +1,2 @@
+"source","type","title","author","date","url","url_date","url_archived","doi"
+"example03","report","Technology Data Generation of Electricity","Danish Energy Agency","2016-08-01","https://ens.dk/media/3273/download","2025-05-06 16:02:04","http://web.archive.org/web/20250506160204/https://ens.dk/media/3273/download",
