+++ technologydata/datasources/example03/technologies.csv
@@ -1,0 +1,3 @@
+"source","technology","detailed_technology","case","region","parameter","year","value","unit",scale,scale_unit,"comment"
+"example01","Solar photovoltaics","Si-HC","example-scenario","DEU","investment",2022,500,"EUR_2022/kW_el","1","MW",
+"example01","Solar photovoltaics","Si-HC","example-project","DEU","investment",2022,350,"EUR_2022/kW_el","1","MW",
