+++ test/test_source_loading.py
@@ -61,3 +61,54 @@ def test_sources_initialization(example_source):
     assert td.Sources([example_source, example_source])
     # Mixed loading of named and loaded sources
     assert td.Sources(["example01", example_source])
+
+
+@pytest.mark.parametrize(
+    "input_date, input_format, output_format, expected_date",
+    [
+        (
+            "20250507105201",
+            "%Y%m%d%H%M%S",
+            "%Y-%m-%d %H:%M:%S",
+            "2025-05-07 10:52:01",
+        ),
+        (
+            "2025-05-07 10:52:01",
+            "%Y-%m-%d %H:%M:%S",
+            "%Y%m%d%H%M%S",
+            "20250507105201",
+        ),
+        (
+            "2025-05-07 10:52:01",
+            "%Y-%m-%d",
+            "%Y%m%d%H%M%S",
+            None,
+        ),
+    ],
+)
+def test_change_datetime_format(
+    input_date, input_format, output_format, expected_date
+) -> None:
+    """Check if the datetime is correctly transformed to a new format."""
+    output_date = Source.change_datetime_format(input_date, input_format, output_format)
+    assert output_date == expected_date
+
+
+@pytest.mark.parametrize(
+    "url, expected",
+    [
+        (
+            "https://ens.dk/media/3273/download",
+            (
+                "http://web.archive.org/web/20250506160204/https://ens.dk/media/3273/download",
+                "2025-05-06 16:02:04",
+                "200",
+            ),
+        ),
+        ("https://ens.dk/media/1/download", None),
+    ],
+)
+def test_is_wayback_snapshot_available(url, expected) -> None:
+    """Check if the example source is available on the Internet Archive Wayback Machine."""
+    output = Source.is_wayback_snapshot_available(url)
+    assert output == expected
